Fix typescript issue
My approach was wrong and caused "npm run build" to fail because
of incompatibility issue with SliderThumb and useFocusRing's focus
prop. I solved the problem initially by modifying the types.d.ts file,
forgetting that it lives in node_modules, and later getting typescript
error when I tried to build in prod. In my defense, react-aria's docs
for the slider component can be confusing, especially for noobs.